Lvl 1 Sound Effects
Added audio for level 1, as well as sound effects for landing on each of the pillars.

I ran into some issues with collisions for all but one of the pillars, I think maybe the mesh hasn't been set up right?

The player is also unable to jump (pc mode) so I haven't fully tested my sound effects.

I tried to add a walking sound effect, but it wasn't working. It might be because of my issue with being able to see the grass texture, maybe part of the terrain just isn't loading right. Will fix l8r